requirements.txt
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.1
python-dotenv==1.0.1
